+++ languages/terraform/default.nix
@@ -75,5 +75,6 @@
           inputs = package.buildInputs;
         });
       };
+      terraform = package;
     };
 }